Bugfix in simulator
Fixed a bug where the elevator would arrive at a floor in the opposite direction of travel when it started moving after being stopped between floors.

Cause:
`prevDir` was used in combination with `prevFloor` to find which floors the elevator was between (prevFloor == 2 and prevDir == Up meant "between floor 2 and 3"). When changing the direction after being stopped between floors, prevDir would also change, thus making the elevator arrive at a floor in the opposite direction it was travelling in.

Fix:
Replace `prevDir` with `departDir`, a variable that is only set when departing a floor (as opposed to when receiving a new MotorDirection). This also removed the need for the `nextFloor` variable (which was only used for generating the display).

Testing:
Tested with the TTK4235 demo

Review:
None

Other:
Recompiled without `-g` for a smaller file and less of the unreadable name-mangled template hell from the stack trace. Ironic, really.